separate order btn from modal
