+++ src/app/[locale]/@modal/(.)order/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import { WriteUsModal } from "@/src/components/shared/WriteUs/WriteUsModal";
+
+const WritUs = () => {
+    return <WriteUsModal />;
+};
+
+export default WritUs;
