--- src/app/[locale]/@modal/(.)write-us/page.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-import { WriteUs } from "@/src/components/shared/WriteUs/WriteUs";
-
-const WritUs = () => {
-    return <WriteUs />;
-};
-
-export default WritUs;
